refactor(web): split the pure half into hash-params.ts, and make "pure" a gate
The pure/glue boundary was two comments inside one file, and the first consumer had already
crossed it: `pages/benchmark/lib/explorer-url.ts` — a pure page lib — imported `hashToSearchParams`
from `hash-adapter.ts` for a one-line `#` strip, and that module's top level imports react + nuqs
and RUNS `unstable_createAdapterProvider(useHashAdapter)`. So `apps/web/test/explorer-url.test.ts`
was building the whole React adapter in vitest's node environment on import. Nothing broke; the
four existing dependency-cruiser rules are all `pages/** <-> shared/**` and cannot see it.

Two files instead of one, same code:

  shared/utils/hash-params.ts   hashToSearchParams, hashUrl, pickKeys. Imports NOTHING.
  shared/utils/hash-adapter.ts  the React/`location`/`history` glue, built on the above.

and a fifth rule, `hash-params-stays-pure` (`from: hash-params.ts`, `to: {}`), which makes the
split checkable instead of declared. Falsified rather than assumed — a temporary
`import { useRef } from 'react'` at the top of hash-params.ts:

  error hash-params-stays-pure: src/shared/utils/hash-params.ts -> .../react@19.2.7/.../index.js
  x 1 dependency violations (1 errors, 0 warnings). 100 modules, 207 dependencies cruised.

restored: `no dependency violations found (100 modules, 206 dependencies cruised)`.

`searchParamsToHash` is deleted with the move: an exported, documented, unit-tested wrapper whose
whole body was `return search.toString()`, with one caller. The reason it exists — why NOT nuqs's
`renderQueryString` — is now a comment on `hashUrl`, where the call is, since `renderQueryString`
is publicly exported from the very module the adapter already imports and is the obvious review
suggestion.

Three more rationales that lived only in commit messages, moved next to the code a cold reviewer
reads:
  - the QueueReset drop (`resetQueues` is internal; we inherit `autoResetQueueOnUpdate ?? true`,
    so only the popstate-specific abort is lost) — on `updateUrl`;
  - a third-party `history.pushState` is invisible until the next event, exactly as in nuqs's own
    React adapter (that is what its opt-in `enableHistorySync` is for, and neither enables it);
    the next adapter write re-syncs rather than clobbering — on `subscribe`;
  - `pickKeys` ALIASES its input on the empty-keys path, as nuqs does — on `pickKeys`.

And the audit the round did not finish, recorded on `getSearchParamsSnapshot`:
`grep -rn 'location\.search' node_modules/nuqs/dist/*.js` finds `dist/index.js:533` as the one
place nuqs still reads the query behind an adapter's back (the others are the react-router adapter
and the opt-in history patch, both unused). It is inert here only because it sits behind an
`onCommittedPathname ||` short-circuit and we supply no `adapter.pathname`, so nuqs compares
`location.pathname`, which never changes under fragment routing. Re-check on a nuqs upgrade.

apps/web/test 16 files / 99 tests, typecheck clean, check-deps 100 modules / 206 dependencies.
