+++ apps/web/src/shared/utils/hash-adapter.ts
@@ -1,57 +1,15 @@
-// The app's URL state lives in the FRAGMENT, not the query string: a playground permalink carries
-// a whole `.s` file, and a query string is part of the HTTP request line (GitHub Pages answers
-// 414 above ~8 KiB), while a fragment is never sent to the server at all.
+// The React glue of the fragment transport: the only part that touches `location` and `history`.
+// The pure helpers it is built on live in hash-params.ts and are unit-tested there.
 //
 // This is a nuqs adapter, so every parser, `useQueryState`, `useQueryStates`, `urlKeys`,
 // `history: 'push'` and the ~20 params of this app keep working untouched — nuqs hands an adapter
 // a `URLSearchParams` and takes one back, and its contract says nothing about where they live.
 // Modelled on nuqs's own `adapters/react`, minus everything a Vite SPA has no use for: no RSC, no
 // SSR snapshot, no provider props, no full-page-navigation branch, no history monkey-patching.
-//
-// The helpers above the glue are pure and tested (apps/web/test/hash-adapter.test.ts); the glue
-// below touches `location`/`history` and is covered by the browser checks instead.
 import { type unstable_AdapterOptions, unstable_createAdapterProvider } from 'nuqs/adapters/custom';
 import { useRef, useSyncExternalStore } from 'react';
 
-// --- the pure halves: strings in, strings out, no `location`, no `history`.
-
-/** `''` | `'#'` | `'#a=1'` | `'a=1'` -> params. Strips exactly one leading `#` — handing one to
- *  `URLSearchParams` would make an entry keyed `'#'` (and `'#a=1'` has no key `a` at all). */
-export function hashToSearchParams(hash: string): URLSearchParams {
-  return new URLSearchParams(hash.startsWith('#') ? hash.slice(1) : hash);
-}
-
-/** params -> the fragment WITHOUT its `#`; `''` for an empty set, which clears the fragment. */
-export function searchParamsToHash(search: URLSearchParams): string {
-  return search.toString();
-}
-
-/** href + params -> the URL to hand to `pushState`. Only the fragment changes; origin, path and
- *  any query survive. Taking the href as an argument is what keeps the write path testable. */
-export function hashUrl(href: string, search: URLSearchParams): string {
-  const url = new URL(href);
-  url.hash = searchParamsToHash(search);
-  return url.toString();
-}
-
-/** Key isolation: a COPY of `search` holding only `keys`, so a hook watching `s` does not
- *  re-render when `tab` moves. nuqs's own `filterSearchParams` is internal (it appears in no
- *  `.d.ts` it ships), so this is a reimplementation of it with `copy` fixed to true.
- *  An empty `keys` means "watch everything", which is what nuqs asks for. */
-export function pickKeys(search: URLSearchParams, keys: string[]): URLSearchParams {
-  if (keys.length === 0) {
-    return search;
-  }
-  const filtered = new URLSearchParams(search);
-  for (const key of Array.from(search.keys())) {
-    if (!keys.includes(key)) {
-      filtered.delete(key);
-    }
-  }
-  return filtered;
-}
-
-// --- the React glue: the only part that touches `location` and `history`.
+import { hashToSearchParams, hashUrl, pickKeys } from './hash-params';
 
 /** `history.pushState` fires NEITHER `popstate` NOR `hashchange`, so this adapter's own writes are
  *  invisible to the DOM. This set is the channel that tells every mounted subscriber about them —
@@ -60,7 +18,13 @@ const listeners = new Set<() => void>();
 
 /** Module-level, therefore stable, which is what stops `useSyncExternalStore` resubscribing on
  *  every render. Three change sources, three disjoint causes: our own writes (the set above),
- *  Back/Forward (`popstate`), and a hand-edited address bar (`hashchange`). */
+ *  Back/Forward (`popstate`), and a hand-edited address bar (`hashchange`).
+ *
+ *  A `history.pushState` by code OUTSIDE this adapter is therefore invisible until the next
+ *  event — nuqs's own React adapter behaves identically, which is what its opt-in
+ *  `enableHistorySync`/`patchHistory` exists for, and neither adapter enables it. Nothing in this
+ *  app writes history out of band, and the next adapter write re-syncs rather than clobbering,
+ *  because `getSearchParamsSnapshot` below re-reads the live hash. */
 function subscribe(onStoreChange: () => void): () => void {
   listeners.add(onStoreChange);
   window.addEventListener('popstate', onStoreChange);
@@ -72,7 +36,12 @@ function subscribe(onStoreChange: () => void): () => void {
   };
 }
 
-/** Closes over nothing, so it is a module-level const rather than a `useMemo`. */
+/** Closes over nothing, so it is a module-level const rather than a `useMemo`.
+ *
+ *  nuqs's own adapter also resets its write queue on `popstate` (its internal `QueueReset`); that
+ *  module is not exported, so this adapter drops it and inherits `autoResetQueueOnUpdate ?? true`
+ *  (nuqs's default), i.e. only the popstate-specific abort is lost. Browser check 2 — one Back
+ *  across a batched Explorer preset — is the observation that no stale queued write lands. */
 function updateUrl(search: URLSearchParams, options: Required<unstable_AdapterOptions>): void {
   const url = hashUrl(window.location.href, search);
   (options.history === 'push' ? history.pushState : history.replaceState).call(history, history.state, '', url);
@@ -88,7 +57,13 @@ function updateUrl(search: URLSearchParams, options: Required<unstable_AdapterOp
  *  `new URLSearchParams(location.search)` (its `getSearchParamsSnapshotFromLocation`), which is
  *  empty under this transport — so without this, a write of one key would drop every OTHER key
  *  from the URL. Found in the browser, not by a test: clicking an Overview aggregate turned
- *  `#view=benchmark` into a URL with no `view` at all. */
+ *  `#view=benchmark` into a URL with no `view` at all.
+ *
+ *  Audited on 2.10.1: `dist/index.js:533` is the ONE place nuqs still reads `location.search`
+ *  behind an adapter's back (`grep -rn 'location\.search' node_modules/nuqs/dist/*.js`, ignoring
+ *  the react-router adapter and the opt-in history patch). It is inert here because it sits in an
+ *  `onCommittedPathname ||` short-circuit and we supply no `adapter.pathname`, so nuqs compares
+ *  `location.pathname`, which never changes under fragment routing. Re-check on a nuqs upgrade. */
 function getSearchParamsSnapshot(): URLSearchParams {
   return hashToSearchParams(window.location.hash);
 }
